requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
cffi==1.14.4
channels==3.0.3
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.4
django-cors==0.1
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
hyperlink==20.0.1
idna==3.1
incremental==17.5.0
joblib==1.0.0
nltk==3.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
pytz==2020.4
regex==2020.11.13
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
textblob==0.15.3
tqdm==4.56.0
Twisted==20.3.0
txaio==20.12.1
urllib3==1.26.2
zope.interface==5.2.0
